+++ CHANGELOG.md
@@ -6,3 +6,4 @@
   - currently just a copy of the overworld for testing purposes
 - Russian translation added by mpustovoi on GitHub
 - fixed weather change packets being sent to all players
+- added office window block